Web: fix save/delete (i64->f64 across wasm boundary) and stale reads
Review of PR #7 found two defects:
- seal_entry/seal_tombstone took Rust i64, which wasm-bindgen expects as
  a JS BigInt, but the page passed a plain Number (Date.now()), so every
  add/edit/delete threw a TypeError before writing. The params are now
  f64 (exact for millisecond timestamps), removing the BigInt boundary.
- API reads used a plain fetch with no cache directive, so the entries
  list could be served stale right after a save. The vault reads now use
  cache: no-store.

+++ web/src/lib.rs
@@ -90,23 +90,33 @@ impl Session {
     /// Encrypt one entry into a record ready to PUT to the server. `data_json`
     /// carries the entry fields; the sealing (and the vault key) stay here, so
     /// the server only ever receives ciphertext. Used for both add and edit.
+    ///
+    /// `modified_ms` is f64 so JavaScript can pass a plain Number (from
+    /// Date.now()); millisecond timestamps are well within f64's exact-integer
+    /// range, and this avoids the i64/BigInt wasm boundary at the call site.
     pub fn seal_entry(
         &self,
         id: &str,
-        modified_ms: i64,
+        modified_ms: f64,
         data_json: &str,
     ) -> Result<String, JsError> {
         let id = Uuid::parse_str(id).map_err(err)?;
         let data: EntryData = serde_json::from_str(data_json).map_err(err)?;
-        let record = self.vault.seal_entry(id, modified_ms, &data).map_err(err)?;
+        let record = self
+            .vault
+            .seal_entry(id, modified_ms as i64, &data)
+            .map_err(err)?;
         serde_json::to_string(&record).map_err(err)
     }
 
     /// Seal a deletion for an entry, ready to PUT to the server. The tombstone
     /// is authenticated under the vault key, so the server cannot forge it.
-    pub fn seal_tombstone(&self, id: &str, modified_ms: i64) -> Result<String, JsError> {
+    pub fn seal_tombstone(&self, id: &str, modified_ms: f64) -> Result<String, JsError> {
         let id = Uuid::parse_str(id).map_err(err)?;
-        let record = self.vault.seal_tombstone(id, modified_ms).map_err(err)?;
+        let record = self
+            .vault
+            .seal_tombstone(id, modified_ms as i64)
+            .map_err(err)?;
         serde_json::to_string(&record).map_err(err)
     }
 }
